+++ backend/SmartPlaylist/Extensions/ListExtensions.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 namespace SmartPlaylist.Extensions
@@ -17,6 +18,18 @@ namespace SmartPlaylist.Extensions
         {
             for (int i = 0; i < Math.Ceiling(source.Count / (double)size); i++)
                 yield return source.Skip(size * i).Take(size);
+        }
+
+        public static IEnumerable<T> DequeueAll<T>(this ConcurrentQueue<T> queue)
+        {
+            List<T> items = new List<T>();
+            while (queue.Count != 0)
+            {
+                T item = default(T);
+                if (queue.TryDequeue(out item))
+                    items.Add(item);
+            }
+            return items;
         }
     }
 }