+++ backend/SmartPlaylist/Plugin.cs
@@ -112,20 +112,20 @@ namespace SmartPlaylist
             {
                 new PluginPageInfo
                 {
-                    Name = "smartplaylist.2.4.0.1.html",
-                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.4.0.1.html",
+                    Name = "smartplaylist.2.4.0.2.html",
+                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.4.0.2.html",
                     EnableInMainMenu = true,
                     MenuIcon = "subscriptions"
                 },
                 new PluginPageInfo
                 {
-                    Name = "smartplaylist.2.4.0.1.css",
-                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.4.0.1.css"
+                    Name = "smartplaylist.2.4.0.2.css",
+                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.4.0.2.css"
                 },
                 new PluginPageInfo
                 {
-                    Name = "smartplaylist.2.4.0.1.js",
-                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.4.0.1.js"
+                    Name = "smartplaylist.2.4.0.2.js",
+                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.4.0.2.js"
                 }
 
             };